formatted, fixed bugs, added functionality
I formatted the code, and fixed the issues that made it so the flags for citizenship and underclassman status wouldn't update. I also fixed the issue where not all of the internships were added to the return value. I also added a parameter that returns the link for the internship. Important note that i commented out the code that sent everything to the database; I did this so testing with the database could be done with a constant json file, to make testing easier.